Remove unused script, use multiprocessing and improve logging system
The 'oxygen_egress.py' file was removed as it is not being used anywhere, possibly an old script or a remnant after refactoring. In 'oxygen_main.py', asyncio's 'gather' was abandoned in favour of Python multiprocessing to enable concurrent execution of 'oxygen_ingress.main()' and 'do_crazy_ai_things.main()', this is necessary to avoid potential deadlocks and utilise multiple cores if possible.

In addition to structural adjustments, logging messages were introduced and updated in 'oxygen_ingress.py' and 'do_crazy_ai_things.py' to make debugging easier and to provide clarity on what the program is doing at any given time. The use of tqdm was discontinued to avoid unnecessary console logs, given the new logging system already provides necessary logs.